email router linked up
