+++ server/routes/employee.router.js
@@ -5,11 +5,6 @@ const {
 const encryptLib = require('../modules/encryption');
 const pool = require('../modules/pool');
 const userStrategy = require('../strategies/user.strategy');
-const sgMail = require('@sendgrid/mail');
-sgMail.setApiKey(process.env.SENDGRID_API_KEY);
-//Moment formating code
-const moment = require('moment');
-var momentPreciseRangePlugin = require("moment-precise-range-plugin");
 
 const router = express.Router();
 
@@ -82,72 +77,5 @@ router.get('/user', rejectUnauthenticated, (req, res) => {
     })
 })
 
-router.get('/email/:id', rejectUnauthenticated, (req, res) => {
-    console.log('timesheet id', req.params.id);
-    const queryText = `SELECT * FROM "timesheet"
-    JOIN "client"
-    ON "timesheet".t_client_id = "client".client_id
-    WHERE "timesheet".timesheet_id = $1;`;
-    const queryValues = [ req.params.id ];
-
-    pool.query(queryText, queryValues)
-    .then( result => {
-        res.send(result.rows[0]);
-        console.log('this is what we get from send email fetch!', result.rows[0]);
-        console.log('this is req.user.email', req.user.email)
-
-        const employeeEmail = req.user.email;
-        const employeeName = req.user.first_name + req.user.last_name
-        const hours =  moment(result.rows[0].clock_out).diff(result.rows[0].clock_in, 'hours');
-        const minutes = moment(result.rows[0].clock_out).diff(result.rows[0].clock_in, 'minutes') % 60;
-
-
-        const employeeMessage = `Timesheet # ${result.rows[0].timesheet_id}
-            ClockIn: ${moment(result.rows[0].clock_in).format('lll')}
-            ClockOut: ${moment(result.rows[0].clock_out).format('lll')}
-            Client: ${result.rows[0].client_first_name} ${result.rows[0].client_last_name }
-            Type of work: ${result.rows[0].work_type}
-            Shift Notes: ${result.rows[0].notes}
-            Time Worked: ${hours}:${minutes}
-            `; //end of employeeMessage
-
-            const adminMessage = `Timesheet # ${result.rows[0].timesheet_id}
-            employee: ${employeeName}
-            ClockIn: ${moment(result.rows[0].clock_in).format('lll')}
-            ClockOut: ${moment(result.rows[0].clock_out).format('lll')}
-            Client: ${result.rows[0].client_first_name} ${result.rows[0].client_last_name }
-            Type of work: ${result.rows[0].work_type}
-            Shift Notes: ${result.rows[0].notes}
-            Time Worked: ${hours}:${minutes}
-            `; //end of adminMessage
-
-        const emails = [ 
-            { //Employee Email Version
-            to: employeeEmail,
-            from: 'ilcsdevs@gmail.com',
-            subject: 'Time Sheet Submitted',
-            text: employeeMessage,
-            // html: '<p>Hello HTML world!</p>',
-            },
-            { //Admin Email Version
-                to: 'ilcsdevs@gmail.com',
-                from: 'ilcsdevs@gmail.com',
-                subject: 'Employee Time Sheet Submitted',
-                text: adminMessage,
-                // html: '<p>Hello HTML world!</p>',
-                },
-            ];
-
-        sgMail
-        .send(emails)
-        .then((response)=> console.log('email sent'))
-        .catch((error)=> console.log(error.message));
-
-        // res.sendStatus(201);
-    }).catch( err => {
-        console.log( err );
-        res.sendStatus(500);
-    })
-})
 
 module.exports = router;
