+++ server/server.js
@@ -19,6 +19,8 @@ const employeeRouter = require('./routes/employee.router')
 const adminTimesheets = require('./routes/adminTimesheets.router')
 const timesheetsEmployeeView = require('./routes/timesheetsEmployeeView.router');
 const adminSingleTimesheet = require('./routes/adminSingleTimesheet.router');
+const email = require('./routes/email.router');
+
 
 // Body parser middleware
 app.use(bodyParser.json());
@@ -43,6 +45,8 @@ app.use('/api/adminTimesheets', adminTimesheets);
 app.use('/api/employeeClockIn', employeeRouter);
 app.use('/api/timesheets/employeeView', timesheetsEmployeeView);
 app.use('/api/admin/timesheet', adminSingleTimesheet);
+app.use('/api/email', email);
+
 
 // Serve static files
 app.use(express.static('build'));
